remove static_cast.

+++ src/graph/graph_apis.cc
@@ -76,16 +76,13 @@ PackedFunc ConvertSubgraphToPackedFunc(const std::vector<NodeFlow>& sg) {
         GraphInterface* gptr = sg[which].graph->Reset();
         GraphHandle ghandle = gptr;
         *rv = ghandle;
-      } else if (which >= static_cast<int>(sg.size()) && which < static_cast<int>(sg.size()) * 2) {
+      } else if (which >= sg.size() && which < sg.size() * 2) {
         *rv = std::move(sg[which - sg.size()].node_mapping);
-      } else if (which >= static_cast<int>(sg.size()) * 2
-                 && which < static_cast<int>(sg.size()) * 3) {
+      } else if (which >= sg.size() * 2 && which < sg.size() * 3) {
         *rv = std::move(sg[which - sg.size() * 2].edge_mapping);
-      } else if (which >= static_cast<int>(sg.size()) * 3
-                 && which < static_cast<int>(sg.size()) * 4) {
+      } else if (which >= sg.size() * 3 && which < sg.size() * 4) {
         *rv = std::move(sg[which - sg.size() * 3].layer_offsets);
-      } else if (which >= static_cast<int>(sg.size()) * 4
-                 && which < static_cast<int>(sg.size()) * 5) {
+      } else if (which >= sg.size() * 4 && which < sg.size() * 5) {
         *rv = std::move(sg[which - sg.size() * 4].flow_offsets);
       } else {
         LOG(FATAL) << "invalid choice";
